requirements.txt
# NBDFinder - Advanced Non-B DNA Analysis Platform
# Core Dependencies for Scientific Analysis

# Web Framework
streamlit>=1.28.0

# Data Science & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.6.0
plotly>=5.15.0
seaborn>=0.12.0
kaleido>=0.2.1
upsetplot>=0.8.0
pycircos>=0.5.2

# File I/O & Data Export
openpyxl>=3.1.0
xlsxwriter>=3.0.0

# Bioinformatics
biopython>=1.81

# Performance Optimization
numba>=0.57.0

# Scientific Computing Support
requests>=2.28.0  # For NCBI API access
